Add files via upload
Choosing scales:

For each of the wavelets, the implementation in pywavelets evaluates the wavelet function for t over the range [wavelet.lower_bound,wavelet.upper_bound] (the default range is [−8,8]). 

scale = 1 corresponds to a case where the extent of the wavelet is (wavelet.upper_bound - wavelet.lower_bound + 1) samples of the digital signal being analysed. Larger scales correspond to the wavelet being stretched. Eg, at scale=10 the wavelet is stretched by a factor of 10, making it sensitive to lower frequencies in the signal.

To relate a given scale to a specific signal frequency, the sampling period of the signal must be known. pywt.scale2frequency() can be used to convert a list of scales to their corresponding frequencies. The proper choice of scales depends on the chosen wavelet, so pywt.scale2frequency() should be used to get an idea of an appropriate range for the signal of interest.

For the cmor, fbsp and shan wavelets, the user can specify a specific a normalized center frequency. A value of 1.0 corresponds to 1/dt where dt is the sampling period. In other words, when analyzing a signal sampled at 100 Hz, a center frequency of 1.0 corresponds to ~100 Hz at scale = 1. This is above the Nyquist rate of 50 Hz, so for this particular wavelet, one would analyze a signal using scales >= 2.

The CWT in PyWavelets is applied to discrete data by convolution with samples of the integral of the wavelet. If scale is too low, this results in a discrete filter that is inadequately sampled and leads to aliasing.